connect four fixed bugs

+++ connectFour/include/logic.h
@@ -2,8 +2,8 @@
 #define LOGIC_H
 
 
-int checkWin(int row, int col, int player, int **grid, short n, short m);
-int drop(int col, int player, int **grid, short n, short m);
+int checkWin(int row, int col, int player, int **grid, int n, int m);
+int drop(int col, int player, int **grid, int n, int m);
 
 
 #endif
